Separate queue admin tests from queue tests
